requirements.txt
fastapi==0.65.1
sqlalchemy==1.4.20
fastapi-login==1.6.0
bcrypt==3.2.0
python-dotenv==0.17.1
email-validator==1.1.2
python-multipart==0.0.5
alembic==1.6.5
python-jose==3.2.0
PyMySQL==1.0.2
sentry-sdk==1.1.0
arq==0.22
aiohttp==3.7.4.post0
emails==0.6
cryptography==3.4.7
Jinja2==3.0.1
types-requests==2.25.2
redisgraph==2.4.0
types-redis==3.5.7
aiofiles==0.7.0
redis==3.5.3
arrow==1.1.1
apscheduler==3.7.0
dvc==2.5.0
grpcio==1.39.0
grpcio-tools==1.39.0
Pillow>=8.2.0
pyyaml>=5.4.1
requests>=2.25.1
supervisor==4.2.2
pynvml==11.0.0
connexion[swagger-ui] >= 2.6.0
connexion==2.6.0
gevent==21.1.2
gunicorn==20.1.0
python_dateutil==2.7.0
setuptools>=21.0.0
ymir-cmd==0.0.2
ymir-proto==0.1.5